requirements.txt
python-dotenv
psycopg2-binary
interval-timer
requests